requirements.txt
asgiref==3.8.1
Django==5.1.6
django-appconf==1.1.0
django-compressor==4.5.1
django-libsass==0.9
django-sass-processor==1.4.1
djangorestframework==3.15.2
libsass==0.23.0
mysqlclient==2.2.7
rcssmin==1.1.2
rest-framework-simplejwt==0.0.2
rjsmin==1.2.2
sqlparse==0.5.3
